Segmentisation modulaire des commandes par classes
Chaque commande est maintenant enveloppée dans une class qui contient les permissions requises pour ces même commandes
Chaque class doit être dans un fichier présent dans le package cmds
Dans un fichier de commande il peut y avoir plusieurs classes et ces fichiers peuvent être regroupés dans des sous-dossier de cmds
