+++ cmds/DefaultCmdClass.py
@@ -1,0 +1,6 @@
+
+class DefaultCmdClass():
+    required_permissions=0
+    dev_boss_only=False
+    home_server_only=False
+    usable_in_dm=False
